requirements.txt
entsoe_py==0.5.8
numpy==1.23.5
pandas==1.5.2
python-dotenv==0.21.0
